# docs/api.openapi.yaml
openapi: 3.1.0
info:
  title: SubnetIQ REST and Streaming API
  version: 1.0.0
  description:
    'Guest mathematics and lookups; Supabase bearer-token persistence; typed assistant SSE. Requests
    accept application/json and are limited to 512 KiB, except project POST/PATCH envelopes (2.25 MiB) with
    a 2 MiB plan cap. Success bodies are returned directly. Failure bodies follow Error. API responses default
    to Cache-Control: no-store. Global quota is 180 requests/minute per IP identity. Production quotas are
    atomic PostgreSQL RPC calls. No credentials means honest guest mode in development; protected operations
    return 503. All ordinary user database access retains the verified user RLS context.'
servers:
  - url: http://localhost:3001
    description: Local API
  - url: /
    description: Current deployment
...
paths:
...
  /api/favorites:
    get:
      operationId: get_favorites
      summary: List favorite tools
      tags:
        - Accounts
      security:
        - supabaseBearer: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Favorite'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Quota exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds until retry.
        '503':
          description: Required credentials, migrations, or service are unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing, invalid, or expired Supabase bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      operationId: post_favorites
      summary: Add or relabel a favorite
      tags:
        - Accounts
      security:
        - supabaseBearer: []
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Favorite'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Quota exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds until retry.
        '503':
          description: Required credentials, migrations, or service are unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing, invalid, or expired Supabase bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tool_id:
                  type: string
                  pattern: ^[a-z][a-z0-9-]{1,80}$
                label:
                  type: string
                  maxLength: 100
                  default: ''
              required:
                - tool_id
              additionalProperties: false
  /api/favorites/{id}:
    delete:
      operationId: delete_favorites_id
      summary: Remove a favorite
      tags:
        - Accounts
      security:
        - supabaseBearer: []
      responses:
        '204':
          description: Deleted or revoked; response body is empty.
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Quota exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds until retry.
        '503':
          description: Required credentials, migrations, or service are unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing, invalid, or expired Supabase bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
        - name: id
          in: path
          required: true
          schema: *id001
  /api/quiz-attempts:
    get:
      operationId: get_quiz_attempts
      summary: List own graded attempts
      tags:
        - Learning
      security:
        - supabaseBearer: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Quota exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds until retry.
        '503':
          description: Required credentials, migrations, or service are unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing, invalid, or expired Supabase bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
        - *id002
        - *id003
    post:
      operationId: post_quiz_attempts
      summary: Grade and record a quiz attempt
      tags:
        - Learning
      security:
        - supabaseBearer: []
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  attempt:
                    type: object
                    additionalProperties: true
                  correct:
                    type: boolean
                  correctIndex:
                    type: integer
                  explanation: *id004
                required:
                  - attempt
                  - correct
                  - correctIndex
                  - explanation
                additionalProperties: false
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Quota exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds until retry.
        '503':
          description: Required credentials, migrations, or service are unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing, invalid, or expired Supabase bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      description:
        The server reconstructs a generated question ID or finds the curated bank entry, validates
        its topic/difficulty and choice index, and grades it. A client-supplied correct flag is rejected. Service-only
        insert maintains derived statistics transactionally.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuizAttemptInput'
  /api/practice/stats:
    get:
      operationId: get_practice_stats
      summary: Read personal topic statistics
      tags:
        - Learning
      security:
        - supabaseBearer: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PracticeStats'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Quota exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds until retry.
        '503':
          description: Required credentials, migrations, or service are unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing, invalid, or expired Supabase bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/history:
    get:
      operationId: get_history
      summary: Read recent cross-feature history
      tags:
        - Accounts
      security:
        - supabaseBearer: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Quota exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds until retry.
        '503':
          description: Required credentials, migrations, or service are unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing, invalid, or expired Supabase bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      description:
        Contains calculations, quizAttempts, conversations and limitPerCategory=100. Use paginated
        individual endpoints for older records.
  /api/profile:
    get:
      operationId: get_profile
      summary: Read own profile and preferences
      tags:
        - Accounts
      security:
        - supabaseBearer: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Quota exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds until retry.
        '503':
          description: Required credentials, migrations, or service are unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing, invalid, or expired Supabase bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      operationId: patch_profile
      summary: Update own profile and preferences
      tags:
        - Accounts
      security:
        - supabaseBearer: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Quota exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds until retry.
        '503':
          description: Required credentials, migrations, or service are unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing, invalid, or expired Supabase bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      description: Avatar must be within the caller UUID prefix. Preferences are limited to 8,000 JSON characters.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                display_name:
                  type: string
                  minLength: 1
                  maxLength: 100
                avatar_path:
                  type:
                    - string
                    - 'null'
                  maxLength: 300
                preferences:
                  type: object
                  additionalProperties: true
              additionalProperties: false
...
  /api/account/export:
    get:
      operationId: get_account_export
      summary: Export own account records
      tags:
        - Accounts
      security:
        - supabaseBearer: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Quota exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds until retry.
        '503':
          description: Required credentials, migrations, or service are unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing, invalid, or expired Supabase bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: Interactive export limit exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      description:
        JSON attachment includes all owned records and identity metadata, excluding share token
        hashes. Bearer tokens/provider keys are never included. Pagination is performed internally. Exports
        above 25 MiB or 50,000 rows per table fail explicitly rather than truncating; use paginated endpoints
        or an operator-assisted export.
...
components:
  securitySchemes:
    supabaseBearer:
      type: http
      scheme: bearer
      bearerFormat: Supabase access JWT
      description:
        A verified Supabase access token; API calls auth.getUser to verify it before creating a
        user-scoped database client.
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: *id004
            message: *id004
            requestId: *id004
            details: {}
          required:
            - code
            - message
          additionalProperties: false
      required:
        - error
      additionalProperties: false
...
    Favorite:
      type: object
      properties:
        id: *id001
        tool_id: *id004
        label: *id004
        created_at:
          type: string
          format: date-time
      required:
        - id
        - tool_id
      additionalProperties: false
...
    QuizAttemptInput:
      type: object
      properties:
        questionId:
          type: string
          minLength: 1
          maxLength: 150
        answerIndex:
          type: integer
          minimum: 0
          maximum: 7
        topic:
          type: string
          minLength: 1
          maxLength: 100
        difficulty:
          type: string
          enum:
            - beginner
            - intermediate
            - advanced
            - exam
        durationMs:
          type: integer
          minimum: 0
          maximum: 86400000
        seed:
          type:
            - string
            - integer
      required:
        - questionId
        - answerIndex
        - topic
        - difficulty
        - durationMs
      additionalProperties: false
    PracticeStats:
      type: object
      properties:
        topic: *id004
        difficulty: *id004
        attempts:
          type: integer
        correct_answers:
          type: integer
        total_duration_ms:
          type: integer
        best_streak:
          type: integer
        current_streak:
          type: integer
        last_practiced_at:
          type: string
          format: date-time
      required:
        - topic
        - difficulty
        - attempts
        - correct_answers
      additionalProperties: false
...